# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MediaTracker",
    "version": "0.2.11",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "ApiHeader": []
    }
  ],
  "paths": {
...
    "/api/progress/by-external-id": {
      "put": {
        "tags": [
          "Progress"
        ],
        "operationId": "addByExternalId",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediaType": {
                    "$ref": "#/components/schemas/MediaType"
                  },
                  "id": {
                    "type": "object",
                    "properties": {
                      "imdbId": {
                        "type": "string",
                        "nullable": true
                      },
                      "tmdbId": {
                        "type": "number",
                        "nullable": true
                      },
                      "audibleId": {
                        "type": "string",
                        "nullable": true
                      },
                      "igdbId": {
                        "type": "number",
                        "nullable": true
                      }
                    }
                  },
                  "seasonNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "episodeNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "action": {
                    "enum": [
                      "paused",
                      "playing",
                      null
                    ],
                    "type": "string",
                    "nullable": true
                  },
                  "progress": {
                    "type": "number",
                    "nullable": true
                  },
                  "duration": {
                    "type": "number",
                    "nullable": true
                  },
                  "device": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "mediaType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/progress/{progressId}": {
      "delete": {
        "tags": [
          "Progress"
        ],
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "progressId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/seen": {
      "put": {
        "tags": [
          "Seen"
        ],
        "operationId": "add",
        "parameters": [
          {
            "name": "mediaItemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "episodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "lastSeenEpisodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "lastSeenAt",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/LastSeenAt"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "notes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/by-external-id": {
      "put": {
        "tags": [
          "Seen"
        ],
        "operationId": "addByExternalId",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediaType": {
                    "$ref": "#/components/schemas/MediaType"
                  },
                  "id": {
                    "type": "object",
                    "properties": {
                      "imdbId": {
                        "type": "string",
                        "nullable": true
                      },
                      "tmdbId": {
                        "type": "number",
                        "nullable": true
                      }
                    }
                  },
                  "seasonNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "episodeNumber": {
                    "type": "number",
                    "nullable": true
                  },
                  "duration": {
                    "type": "number",
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "mediaType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/{seenId}": {
      "put": {
        "tags": [
          "Seen"
        ],
        "operationId": "updateById",
        "parameters": [
          {
            "name": "seenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "nullable": true
                    },
                    "date": {
                      "type": "number",
                      "nullable": true
                    },
                    "mediaItemId": {
                      "type": "number",
                      "nullable": true
                    },
                    "episodeId": {
                      "type": "number",
                      "nullable": true
                    },
                    "userId": {
                      "type": "number",
                      "nullable": true
                    },
                    "type": {
                      "enum": [
                        "progress",
                        "seen",
                        null
                      ],
                      "type": "string",
                      "nullable": true
                    },
                    "progress": {
                      "type": "number",
                      "nullable": true
                    },
                    "duration": {
                      "type": "number",
                      "nullable": true
                    },
                    "action": {
                      "enum": [
                        "paused",
                        "playing",
                        null
                      ],
                      "type": "string",
                      "nullable": true
                    },
                    "notes": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Seen"
        ],
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "seenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/seen/": {
      "delete": {
        "tags": [
          "Seen"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "mediaItemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "seasonId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "episodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "ApiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "MediaType": {
        "enum": [
          "audiobook",
          "book",
          "movie",
          "tv",
          "video_game"
        ],
        "type": "string"
      },
...
      "LastSeenAt": {
        "enum": [
          "custom_date",
          "now",
          "release_date",
          "unknown"
        ],
        "type": "string"
      },
...
    }
  }
}